% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_nc}
\alias{download_nc}
\title{Download NetCDF Files}
\usage{
download_nc(nc_links, file_base, out_dir)
}
\arguments{
\item{nc_links}{Character vector. A list of NetCDF file links to download.}

\item{file_base}{Character. The base URL for the files to be downloaded.}

\item{out_dir}{Character. The output directory where the downloaded files will be saved.}
}
\value{
None. The function downloads files to the specified output directory.
}
\description{
This function downloads NetCDF files from a specified list of links.
}
\examples{
download_nc(nc_links=c("file1.nc", "file2.nc"), file_base="https://example.com/files/", out_dir="~/data/")
}
