% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndExtract.R
\name{integrateShortWave}
\alias{integrateShortWave}
\title{Integrate daily accumulation of shortwave radiation}
\usage{
integrateShortWave(shortwave.mx, rowNum, endTime, startTime = 0)
}
\arguments{
\item{shortwave.mx}{Matrix of short_wave variable [elem, hour]}

\item{rowNum}{Row number of shortwave.mx}

\item{endTime}{Upper integration time limit (max: 23)}

\item{startTime}{Lower integration time limit (min: 0S)}
}
\value{
Vector of values, length nrow(shortwave.mx)
}
\description{
Integrate daily accumulation of shortwave radiation
}
