% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndExtract.R
\name{loadHydroVars}
\alias{loadHydroVars}
\title{Load hydrodynamic variables from WeStCOMS meshes}
\usage{
loadHydroVars(
  sampling.df,
  westcoms.dir,
  vars,
  daySummaryFn = NULL,
  depthSummaryFn = NULL,
  sep = "/",
  cores = 1,
  progress = T
)
}
\arguments{
\item{sampling.df}{Dataframe with a row for each sample. Columns must include
`site.id`, `obs.id`, `date` (YYYYMMDD), `hour`, `depth`, `grid`,
`site.elem`, `trinode_1`, `trinode_2`, and `trinode_3`, which can be added
through a spatial join with the mesh files (see vignette)}

\item{westcoms.dir}{Character vector with directories for WeStCOMS v1 and
WeStCOMS v2 hydrodynamic files}

\item{vars}{Character vector of hydrodynamic variables to extract}

\item{daySummaryFn}{Vector of functions to summarise by day, one per var; if
`NULL` (default), single hour is extracted}

\item{depthSummaryFn}{Vector of functions to summarise across depth, one per
var; if `NULL` (default), single hour is extracted}

\item{sep}{Directory separation character (Windows: '\\', Unix: '/')}

\item{cores}{Number of cores for extracting in parallel; default is 1}

\item{progress}{Logical: Show progress bar?}
}
\value{
dataframe with site.id, date, and hydrodynamic variables
}
\description{
Load hydrodynamic variables from WeStCOMS meshes
}
