% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndExtract.R
\name{loadMesh}
\alias{loadMesh}
\title{Load WeStCOMS mesh}
\usage{
loadMesh(mesh.f, type = "sf")
}
\arguments{
\item{mesh.f}{Mesh file name}

\item{type}{Output type: `sf` (default) or `nc`}
}
\value{
Mesh object (ncdf or sf)
}
\description{
Load WeStCOMS mesh
}
