% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndExtract.R
\name{meanOfNodes}
\alias{meanOfNodes}
\title{Find mean of trinodes}
\usage{
meanOfNodes(nc.ar, node.mx)
}
\arguments{
\item{nc.ar}{Hydrodynamic variable array}

\item{node.mx}{Trinode index matrix}
}
\value{
Mean value with same dimensions as nc.ar
}
\description{
Find the element mean, not weighting by distance (yet)
}
