% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndExtract.R
\name{q90}
\alias{q90}
\title{Find 90th quantile}
\usage{
q90(x)
}
\arguments{
\item{x}{Vector}
}
\value{
vector of quantiles
}
\description{
Same as quantile(x, probs=0.9), but works with just a single argument
}
