% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndExtract.R
\name{valueOfElement}
\alias{valueOfElement}
\title{Find value of element}
\usage{
valueOfElement(nc.ar, elem.id)
}
\arguments{
\item{nc.ar}{Hydrodynamic variable array}

\item{elem.id}{Element id}
}
\value{
Value with same dimensions as nc.ar
}
\description{
Find the element value, not weighting by distance (yet)
}
